Finished Ghost PathFinding

+++ src/main/java/Player.java
@@ -24,11 +24,11 @@ public class Player implements Unit {
 
     }
 
-    public int getX() {
+    public float getX() {
         return this.x;
     }
 
-    public int getY() {
+    public float getY() {
         return this.y;
     }
 }
